Fix for Node.js 11 (#448)
The `sortOrder()` function appears to make an assumption about custom
sort functions that isn't true in the V8 shipped with Node.js 11. As a
result, tests fail in Node.js 11, and probably code in the wild fails
too. This fixes the function to be compatible with Node.js 11.

This will also mean that Travis should now start passing again.